+++ app/src/main/java/com/f2h/f2h_buyer/screens/group/pre_order/TableComponent.kt
@@ -4,6 +4,6 @@ package com.f2h.f2h_buyer.screens.group.pre_order
 data class TableComponent (
     var id: Long = 0L,
     var date: String = "",
-    var quantity: Float = 0F,
+    var quantity: Double = 0.0,
     var uom: String = ""
 )
